Add contribution guidelines and a code of conduct  (#7)
* Add a Code of Conduct

Use the Contributor Covenant Code of Conduct, as recommended by [the guidelines to create a awesome list](https://github.com/sindresorhus/awesome/blob/master/create-list.md)

* Add contribution guidelines

As needed by the guidelines to create a awesome list, add some contribution guidelines. They are based on the one of awesome-macos. It's just a start, and more guidelines might have to be added, especially in what kind of app goes in the list.

* Add the rules for proposing applications and extensions

* Forgot some punctuation